feat(sidebar): move group-by control into view options menu
Consolidates the standalone Group by segmented control into the view
options dropdown alongside Sort by and Show properties, and opens the
menu to the right of the trigger so it no longer covers the worktree
list.
