+++ src/renderer/src/components/sidebar/index.tsx
@@ -5,7 +5,6 @@ import { useSidebarResize } from '@/hooks/useSidebarResize'
 import SidebarHeader from './SidebarHeader'
 import SidebarNav from './SidebarNav'
 import SearchBar from './SearchBar'
-import GroupControls from './GroupControls'
 import WorktreeList from './WorktreeList'
 import SidebarToolbar from './SidebarToolbar'
 import WorktreeMetaDialog from './WorktreeMetaDialog'
@@ -51,7 +50,6 @@ function Sidebar(): React.JSX.Element {
         <SidebarNav />
         <SidebarHeader />
         <SearchBar />
-        <GroupControls />
 
         {/* Virtualized scrollable list */}
         <WorktreeList />
